Diseño

+++ app/Http/Controllers/Admin/GrupoMenuController.php
@@ -98,11 +98,13 @@ class GrupoMenuController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        GrupoMenu::destroy($id);
-        return redirect()
-            ->route('grupomenu')
-            ->with('success', 'Eliminado Correctamente');
+        if ($request->ajax()) {
+            GrupoMenu::destroy($id);
+            return response()->json(['mensaje' => 'ok']);
+        } else {
+            abort(404);
+        }
     }
 }
